move functions to new utils file
